+++ python/GafferSceneUI/LightEditor.py
@@ -106,7 +106,13 @@ class LightEditor( GafferUI.NodeSetEditor ) :
 
 			self.__pathListing = GafferUI.PathListingWidget(
 				Gaffer.DictPath( {}, "/" ), # Temp till we make a ScenePath
-				columns = [ _GafferSceneUI._LightEditorLocationNameColumn() ],
+				columns = [
+					_GafferSceneUI._LightEditorLocationNameColumn(),
+					_GafferSceneUI._LightEditorMuteColumn(
+						self.__settingsNode["in"],
+						self.__settingsNode["editScope"]
+					),
+				],
 				selectionMode = GafferUI.PathListingWidget.SelectionMode.Cells,
 				displayMode = GafferUI.PathListingWidget.DisplayMode.Tree,
 				horizontalScrollMode = GafferUI.ScrollMode.Automatic
@@ -241,7 +247,8 @@ class LightEditor( GafferUI.NodeSetEditor ) :
 				sectionColumns += [ c( self.__settingsNode["in"], self.__settingsNode["editScope"] ) for c in section.values() ]
 
 		nameColumn = self.__pathListing.getColumns()[0]
-		self.__pathListing.setColumns( [ nameColumn ] + sectionColumns )
+		muteColumn = self.__pathListing.getColumns()[1]
+		self.__pathListing.setColumns( [ nameColumn, muteColumn ] + sectionColumns )
 
 	def __settingsPlugSet( self, plug ) :
 
@@ -309,41 +316,50 @@ class LightEditor( GafferUI.NodeSetEditor ) :
 
 		return False
 
-	def __editSelectedCells( self, pathListing ) :
+	def __editSelectedCells( self, pathListing, quickBoolean = True ) :
 
-		selection = pathListing.getSelection()
-
-		columns = pathListing.getColumns()
-
+		# A dictionary of the form :
+		# { inspector : { path1 : inspection, path2 : inspection, ... }, ... }
+		inspectors = {}
 		inspections = []
 
 		with Gaffer.Context( self.getContext() ) as context :
-			for i in range( 0, len( columns ) ) :
-				column = columns[ i ]
+			for selection, column in zip( pathListing.getSelection(), pathListing.getColumns() ) :
 				if not isinstance( column, _GafferSceneUI._LightEditorInspectorColumn ) :
 					continue
-				for path in selection[i].paths() :
-					context["scene:path"] = GafferScene.ScenePlug.stringToPath( path )
+				for pathString in selection.paths() :
+					path = GafferScene.ScenePlug.stringToPath( pathString )
+					context["scene:path"] = path
 					inspection = column.inspector().inspect()
 
 					if inspection is not None :
+						inspectors.setdefault( column.inspector(), {} )[path] = inspection
 						inspections.append( inspection )
 
-		if len( inspections ) == 0 :
+		if len( inspectors ) == 0 :
+			with GafferUI.PopupWindow() as self.__popup :
+				with GafferUI.ListContainer( GafferUI.ListContainer.Orientation.Horizontal, spacing = 4 ) :
+					GafferUI.Image( "warningSmall.png" )
+					GafferUI.Label( "<h4>The selected cells cannot be edited in the current Edit Scope</h4>" )
+
+			self.__popup.popup()
+
 			return
 
 		nonEditable = [ i for i in inspections if not i.editable() ]
 
 		if len( nonEditable ) == 0 :
-			edits = [ i.acquireEdit() for i in inspections ]
-			warnings = "\n".join( [ i.editWarning() for i in inspections if i.editWarning() != "" ] )
+			if not quickBoolean or not self.__toggleBoolean( inspectors, inspections ) :
+				edits = [ i.acquireEdit() for i in inspections ]
+				warnings = "\n".join( [ i.editWarning() for i in inspections if i.editWarning() != "" ] )
+				# The plugs are either not boolean, boolean with mixed values,
+				# or attributes that don't exist and are not boolean. Show the popup.
+				self.__popup = GafferUI.PlugPopup( edits, warning = warnings )
 
-			self.__popup = GafferUI.PlugPopup( edits, warning = warnings )
+				if isinstance( self.__popup.plugValueWidget(), GafferUI.TweakPlugValueWidget ) :
+					self.__popup.plugValueWidget().setNameVisible( False )
 
-			if isinstance( self.__popup.plugValueWidget(), GafferUI.TweakPlugValueWidget ) :
-				self.__popup.plugValueWidget().setNameVisible( False )
-
-			self.__popup.popup()
+				self.__popup.popup()
 
 		else :
 
@@ -353,6 +369,54 @@ class LightEditor( GafferUI.NodeSetEditor ) :
 					GafferUI.Label( "<h4>{}</h4>".format( nonEditable[0].nonEditableReason() ) )
 
 			self.__popup.popup()
+
+	def __toggleBoolean( self, inspectors, inspections ) :
+
+		plugs = [ i.acquireEdit() for i in inspections ]
+		# Make sure all the plugs are either a BoolPlug or contain a BoolPlug
+		if not all (
+			(
+				isinstance( p, ( Gaffer.TweakPlug, Gaffer.NameValuePlug ) ) and
+				isinstance( p["value"], Gaffer.BoolPlug )
+			) or isinstance( p, Gaffer.BoolPlug ) for p in plugs
+		) :
+			return False
+
+		currentValues = []
+
+		# Use a single new value for all plugs.
+		# First we need to find out what the new value would be for each plug in isolation.
+		for inspector, pathInspections in inspectors.items() :
+			for path, inspection in pathInspections.items() :
+				currentValue = inspection.value().value if inspection.value() is not None else None
+
+				if isinstance( inspector, GafferSceneUI.Private.AttributeInspector ) :
+					fullAttributes = self.__settingsNode["in"].fullAttributes( path[:-1] )
+					parentValueData = fullAttributes.get( inspector.name(), None )
+					parentValue = parentValueData.value if parentValueData is not None else False
+
+					currentValues.append( currentValue if currentValue is not None else parentValue )
+				else :
+					currentValues.append( currentValue )
+
+		# Now set the value for all plugs, defaulting to `True` if they are not
+		# currently all the same.
+		newValue = not sole( currentValues )
+
+		with Gaffer.UndoScope( self.scriptNode() ) :
+			for inspector, pathInspections in inspectors.items() :
+				for path, inspection in pathInspections.items() :
+					plug = inspection.acquireEdit()
+
+					if isinstance( plug, ( Gaffer.TweakPlug, Gaffer.NameValuePlug ) ) :
+						plug["value"].setValue( newValue )
+						plug["enabled"].setValue( True )
+						if isinstance( plug, Gaffer.TweakPlug ) :
+							plug["mode"].setValue( Gaffer.TweakPlug.Mode.Create )
+					else :
+						plug.setValue( newValue )
+
+		return True
 
 	def __buttonPress( self, pathListing, event ) :
 
@@ -434,6 +498,12 @@ class LightEditor( GafferUI.NodeSetEditor ) :
 				"Show History...",
 				{
 					"command" : Gaffer.WeakMethod( self.__showHistory )
+				}
+			)
+			menuDefinition.append(
+				"Edit...",
+				{
+					"command" : functools.partial( self.__editSelectedCells, pathListing, False )
 				}
 			)